Rename simple_pricing to customer_group_pricing
Refs SHOOP-2101
