+++ doc/addons.rst
@@ -29,7 +29,7 @@ For instance, take a bare-bones Shoop core installation.
         'django.contrib.messages',
         'django.contrib.staticfiles',
         'shoop.core',
-        'shoop.simple_pricing',
+        'shoop.customer_group_pricing',
         'shoop.simple_supplier',
         'shoop.default_tax',
         'shoop.admin',
@@ -62,7 +62,7 @@ Putting this all together,
         'django.contrib.messages',
         'django.contrib.staticfiles',
         'shoop.core',
-        'shoop.simple_pricing',
+        'shoop.customer_group_pricing',
         'shoop.simple_supplier',
         'shoop.default_tax',
         'shoop.admin',
