+++ doc/api/shoop.customer_group_pricing.rst
@@ -1,0 +1,46 @@
+shoop.customer_group_pricing package
+====================================
+
+Submodules
+----------
+
+shoop.customer_group_pricing.admin_form_part module
+---------------------------------------------------
+
+.. automodule:: shoop.customer_group_pricing.admin_form_part
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+shoop.customer_group_pricing.api module
+---------------------------------------
+
+.. automodule:: shoop.customer_group_pricing.api
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+shoop.customer_group_pricing.models module
+------------------------------------------
+
+.. automodule:: shoop.customer_group_pricing.models
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+shoop.customer_group_pricing.module module
+------------------------------------------
+
+.. automodule:: shoop.customer_group_pricing.module
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: shoop.customer_group_pricing
+    :members:
+    :undoc-members:
+    :show-inheritance:
