+++ doc/api/shoop.rst
@@ -12,12 +12,12 @@ Subpackages
     shoop.apps
     shoop.campaigns
     shoop.core
+    shoop.customer_group_pricing
     shoop.default_tax
     shoop.discount_pricing
     shoop.front
     shoop.notify
     shoop.simple_cms
-    shoop.simple_pricing
     shoop.simple_supplier
     shoop.testing
     shoop.themes
