--- doc/api/shoop.simple_pricing.rst
@@ -1,46 +1,0 @@
-shoop.simple_pricing package
-============================
-
-Submodules
-----------
-
-shoop.simple_pricing.admin_form_part module
--------------------------------------------
-
-.. automodule:: shoop.simple_pricing.admin_form_part
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-shoop.simple_pricing.api module
--------------------------------
-
-.. automodule:: shoop.simple_pricing.api
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-shoop.simple_pricing.models module
-----------------------------------
-
-.. automodule:: shoop.simple_pricing.models
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-shoop.simple_pricing.module module
-----------------------------------
-
-.. automodule:: shoop.simple_pricing.module
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: shoop.simple_pricing
-    :members:
-    :undoc-members:
-    :show-inheritance:
