+++ doc/api/shoop.simple_supplier.admin_module.rst
@@ -1,0 +1,30 @@
+shoop.simple_supplier.admin_module package
+==========================================
+
+Submodules
+----------
+
+shoop.simple_supplier.admin_module.forms module
+-----------------------------------------------
+
+.. automodule:: shoop.simple_supplier.admin_module.forms
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+shoop.simple_supplier.admin_module.views module
+-----------------------------------------------
+
+.. automodule:: shoop.simple_supplier.admin_module.views
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: shoop.simple_supplier.admin_module
+    :members:
+    :undoc-members:
+    :show-inheritance:
