+++ shoop/customer_group_pricing/locale/fi_FI/LC_MESSAGES/django.po
@@ -1,0 +1,70 @@
+# SOME DESCRIPTIVE TITLE.
+# Copyright (C) 2015 THE PACKAGE'S COPYRIGHT HOLDER
+# This file is distributed under the same license as the PACKAGE package.
+#
+# Translators:
+# Aarni Koskela, 2015
+# Antti Mäki <antti@anttimaki.com>, 2015
+# Eetu Häivälä <eetu.haivala@anders.fi>, 2015
+# Janne Tammi <janne.tammi@shoop.io>, 2015
+# Joni Latokylä <joni.latokyla@andersinnovations.com>, 2015
+# Juha Kujala <juha.kujala@shoop.io>, 2015
+# Juha Louhiranta <charni@gmail.com>, 2015
+# Rauli Laine <rauli.laine@iki.fi>, 2015
+# Shoop Admin <admin@shoop.io>, 2015
+# Teemu Selkälä <teemu.selkala@anders.fi>, 2015
+# tomijarvi <tomi@jarvi.pro>, 2015
+# Tuomas Suutari <tuomas.suutari@anders.fi>, 2015
+msgid ""
+msgstr ""
+"Project-Id-Version: Shoop\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-02-01 22:46+0000\n"
+"PO-Revision-Date: 2015-12-08 08:38+0000\n"
+"Last-Translator: Shoop Admin <admin@shoop.io>\n"
+"Language-Team: Finnish (Finland) (http://www.transifex.com/shoop/shoop/"
+"language/fi_FI/)\n"
+"Language: fi_FI\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Generated-By: Babel 2.1.1\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+
+#, fuzzy
+#| msgid "Customer Group Pricing"
+msgid "Shoop Customer Group Pricing"
+msgstr "Asiakasryhmähinnoittelu"
+
+#, python-format
+msgid "Price (%(shop)s/%(group)s)"
+msgstr "Hinta (%(shop)s/%(group)s)"
+
+#, fuzzy
+#| msgid "product price"
+msgid "product"
+msgstr "tuotehinta"
+
+msgid "shop"
+msgstr ""
+
+msgid "contact group"
+msgstr ""
+
+msgid "price"
+msgstr ""
+
+msgid "product price"
+msgstr "tuotehinta"
+
+msgid "product prices"
+msgstr "tuotehinnat"
+
+msgid "Customer Group Pricing"
+msgstr "Asiakasryhmähinnoittelu"
+
+msgid "Default price"
+msgstr "Oletushinta"
+
+msgid "Other shops"
+msgstr "Muut kaupat"
