+++ shoop/customer_group_pricing/locale/ja_JP/LC_MESSAGES/django.po
@@ -1,0 +1,57 @@
+# SOME DESCRIPTIVE TITLE.
+# Copyright (C) 2015 THE PACKAGE'S COPYRIGHT HOLDER
+# This file is distributed under the same license as the PACKAGE package.
+#
+# Translators:
+# Yuki Miyagi <yuki.miyagi@anders.fi>, 2015
+msgid ""
+msgstr ""
+"Project-Id-Version: Shoop\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-02-01 22:46+0000\n"
+"PO-Revision-Date: 2015-12-08 08:37+0000\n"
+"Last-Translator: Shoop Admin <admin@shoop.io>\n"
+"Language-Team: Japanese (Japan) (http://www.transifex.com/shoop/shoop/"
+"language/ja_JP/)\n"
+"Language: ja_JP\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Generated-By: Babel 2.1.1\n"
+"Plural-Forms: nplurals=1; plural=0;\n"
+
+msgid "Shoop Customer Group Pricing"
+msgstr ""
+
+#, python-format
+msgid "Price (%(shop)s/%(group)s)"
+msgstr ""
+
+#, fuzzy
+#| msgid "product price"
+msgid "product"
+msgstr "商品価格"
+
+msgid "shop"
+msgstr ""
+
+msgid "contact group"
+msgstr ""
+
+msgid "price"
+msgstr ""
+
+msgid "product price"
+msgstr "商品価格"
+
+msgid "product prices"
+msgstr "商品価格"
+
+msgid "Customer Group Pricing"
+msgstr ""
+
+msgid "Default price"
+msgstr ""
+
+msgid "Other shops"
+msgstr ""
