--- shoop/simple_pricing/locale/en/LC_MESSAGES/django.po
@@ -1,54 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) 2015 THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the PACKAGE package.
-# FIRST AUTHOR <EMAIL@ADDRESS>, 2015.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PACKAGE VERSION\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-01-20 13:16+0000\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: en <LL@li.org>\n"
-"Language: en\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"Generated-By: Babel 2.1.1\n"
-
-#, python-format
-msgid "Price (%(shop)s/%(group)s)"
-msgstr ""
-
-msgid "product"
-msgstr ""
-
-msgid "shop"
-msgstr ""
-
-msgid "contact group"
-msgstr ""
-
-msgid "price"
-msgstr ""
-
-msgid "product price"
-msgstr ""
-
-msgid "product prices"
-msgstr ""
-
-msgid "Simple Pricing"
-msgstr ""
-
-msgid "Customer Group Pricing"
-msgstr ""
-
-msgid "Default price"
-msgstr ""
-
-msgid "Other shops"
-msgstr ""
