--- shoop/simple_pricing/locale/zh_Hans/LC_MESSAGES/django.po
@@ -1,57 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) 2015 THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the PACKAGE package.
-#
-# Translators:
-# Jessica Wang <jessica.wang@shoop.io>, 2015
-msgid ""
-msgstr ""
-"Project-Id-Version: Shoop\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-01-21 07:22+0000\n"
-"PO-Revision-Date: 2015-12-10 09:19+0000\n"
-"Last-Translator: Jessica Wang <jessica.wang@shoop.io>\n"
-"Language-Team: Chinese (Mandarin) (http://www.transifex.com/shoop/shoop/"
-"language/cmn/)\n"
-"Language: cmn\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel 2.1.1\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-
-#, python-format
-msgid "Price (%(shop)s/%(group)s)"
-msgstr "价格 (%(shop)s/%(group)s)"
-
-#, fuzzy
-#| msgid "product price"
-msgid "product"
-msgstr "产品价格"
-
-msgid "shop"
-msgstr ""
-
-msgid "contact group"
-msgstr ""
-
-msgid "price"
-msgstr ""
-
-msgid "product price"
-msgstr "产品价格"
-
-msgid "product prices"
-msgstr "产品价格"
-
-msgid "Simple Pricing"
-msgstr "简单定价"
-
-msgid "Customer Group Pricing"
-msgstr "顾客群组定价"
-
-msgid "Default price"
-msgstr "默认价格"
-
-msgid "Other shops"
-msgstr "其它店铺"
