--- shoop/simple_pricing/templates/shoop/admin/simple_pricing/form_part.jinja
@@ -1,24 +1,0 @@
-{% from "shoop/admin/macros/general.jinja" import content_block with context %}
-{% set sp_form = form["simple_pricing"] %}
-{% call content_block(_("Customer Group Pricing"), "fa-money") %}
-    <div class="table-responsive">
-        <table class="table table-bordered table-condensed table-striped">
-            <tbody>
-            <tr>
-                <th></th>
-                {% for group in sp_form.groups %}
-                    <th scope="col">{% if group %}{{ group }}{% else %}<i>{% trans %}Default price{% endtrans %}</i>{% endif %}</th>
-                {% endfor %}
-            </tr>
-            {% for shop in sp_form.shops %}
-                <tr>
-                    <th scope="row">{% if shop %}{{ shop }}{% else %}<i>{% trans %}Other shops{% endtrans %}</i>{% endif %}</th>
-                    {% for group in sp_form.groups %}
-                        <td>{{ bs3.field(sp_form.get_shop_group_field(shop, group), set_placeholder=False, render_label=False) }}</td>
-                    {% endfor %}
-                </tr>
-            {% endfor %}
-            </tbody>
-        </table>
-    </div>
-{% endcall %}
